refactor scheduler components into smaller parts
